make codecov happy
