+++ packages/remax-cli/src/__tests__/fixtures/alipay/expected/pages/class.axml
@@ -320,49 +320,6 @@
   </movable-view>
 </template>
   
-  <template name="REMAX_TPL_1_native-component-0">
-  <native-component-0 ref="{{item.props['__ref']}}">
-    
-    <block a:for="{{item.children}}" key="{{item.id}}">
-      <block a:if="{{item.props['slot']}}">
-        <view slot="{{item.props['slot']}}" 
-          animation="{{item.props['animation']}}" 
-          class="{{item.props['class']}}" 
-          disable-scroll="{{item.props['disable-scroll']}}" 
-          hidden="{{item.props['hidden']}}" 
-          hover-class="{{item.props['hover-class']}}" 
-          hover-start-time="{{item.props['hover-start-time']}}" 
-          hover-stay-time="{{item.props['hover-stay-time']}}" 
-          hover-stop-propagation="{{item.props['hover-stop-propagation']}}" 
-          id="{{item.props['id']}}" 
-          onAnimationEnd="{{item.props['onAnimationEnd']}}" 
-          onAnimationIteration="{{item.props['onAnimationIteration']}}" 
-          onAnimationStart="{{item.props['onAnimationStart']}}" 
-          onAppear="{{item.props['onAppear']}}" 
-          onDisappear="{{item.props['onDisappear']}}" 
-          onFirstAppear="{{item.props['onFirstAppear']}}" 
-          onLongTap="{{item.props['onLongTap']}}" 
-          onTap="{{item.props['onTap']}}" 
-          onTouchCancel="{{item.props['onTouchCancel']}}" 
-          onTouchEnd="{{item.props['onTouchEnd']}}" 
-          onTouchMove="{{item.props['onTouchMove']}}" 
-          onTouchStart="{{item.props['onTouchStart']}}" 
-          onTransitionEnd="{{item.props['onTransitionEnd']}}" 
-          style="{{item.props['style']}}" 
-        >
-          <block a:for="{{item.children}}" key="{{item.id}}">
-            <template is="REMAX_TPL_1_CONTAINER" data="{{item: item}}" />
-          </block>
-        </view>
-      </block>
-      <block a:else>
-        <template is="REMAX_TPL_1_CONTAINER" data="{{item: item}}" />
-      </block>
-    </block>
-    
-  </native-component-0>
-</template>
-  
   <template name="REMAX_TPL_1_navigator">
   <navigator 
     class="{{item.props['class']}}"
@@ -693,6 +650,7 @@
   <view 
     animation="{{item.props['animation']}}"
     class="{{item.props['class']}}"
+    data-foo="{{item.props['data-foo']}}"
     disable-scroll="{{item.props['disable-scroll']}}"
     hidden="{{item.props['hidden']}}"
     hover-class="{{item.props['hover-class']}}"
