+++ packages/remax-cli/src/__tests__/fixtures/alipay/expected/pages/class.json
@@ -1,0 +1,3 @@
+{
+  "usingComponents": {}
+}
