+++ packages/remax-cli/src/__tests__/fixtures/alipay/src/app.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-  pages: ['pages/index', 'pages/dont-exist/index'],
+  pages: ['pages/index', 'pages/dont-exist/index', 'pages/class'],
   window: {
     defaultTitle: 'Alipay App',
     titleBarColor: '#323239',
