+++ packages/remax-cli/src/__tests__/fixtures/alipay/src/pages/class.js
@@ -1,0 +1,9 @@
+import * as React from 'react';
+import { View } from 'remax/alipay';
+
+// 测试 page 组件为 class component 的情况
+export default class RemaxPage {
+  render() {
+    return <View>remax class component page test</View>;
+  }
+}
