+++ packages/remax-cli/src/__tests__/fixtures/native/expected/components/native-component/index.axml
@@ -1,0 +1,1 @@
+<view>a</view>
