+++ packages/remax-cli/src/__tests__/fixtures/native/expected/components/native-component/index.js
@@ -1,0 +1,8 @@
+'use strict';
+
+Object.defineProperty(exports, '__esModule', { value: true });
+
+Component({});
+var index = {};
+
+exports.default = index;
