+++ packages/remax-cli/src/__tests__/fixtures/native/expected/pages/remax/index.js
@@ -7,8 +7,11 @@ require('scheduler');
 var remaxVendor = require('../../remaxVendor.js');
 var React = require('react');
 
+var NativeComponent = remaxVendor.createNativeComponent('native-component-0');
+
+
 function _ref(props) {
-  return React.createElement(remaxVendor.View, null, "alipay view");
+  return React.createElement(remaxVendor.View, null, React.createElement(NativeComponent, null), "alipay view");
 }
 
 var index = Page(remaxVendor.createPageConfig(_ref));
