+++ packages/remax-cli/src/__tests__/fixtures/native/expected/pages/remax/index.json
@@ -1,3 +1,5 @@
 {
-  "usingComponents": {}
+  "usingComponents": {
+    "native-component-0": "../../components/native-component/index"
+  }
 }
