+++ packages/remax-cli/src/__tests__/fixtures/native/expected/remaxVendor.js
@@ -1246,6 +1246,87 @@ var hostConfig = {
 
 var ReactReconcilerInst = ReactReconciler(hostConfig);
 
+function getPublicRootInstance(container) {
+  var containerFiber = container.current;
+
+  if (!containerFiber.child) {
+    return null;
+  }
+
+  return containerFiber.child.stateNode;
+}
+
+function render(rootElement, container) {
+  // Create a root Container if it doesnt exist
+  if (!container._rootContainer) {
+    container._rootContainer = ReactReconcilerInst.createContainer(container, false, false);
+  }
+
+  ReactReconcilerInst.updateContainer(rootElement, container._rootContainer, null, function () {// ignore
+  });
+  return getPublicRootInstance(container._rootContainer);
+}
+
+var AppContainer =
+/** @class */
+function () {
+  function AppContainer(context) {
+    this.updateQueue = [];
+    this.context = context;
+    this.root = new VNode({
+      id: generate(),
+      type: 'root',
+      container: this
+    });
+    this.root.mounted = true;
+  }
+
+  AppContainer.prototype.requestUpdate = function (path, start, deleteCount) {
+    var items = [];
+
+    for (var _i = 3; _i < arguments.length; _i++) {
+      items[_i - 3] = arguments[_i];
+    } // ignore
+
+  };
+
+  AppContainer.prototype.createCallback = function (name, fn) {
+    this.context[name] = fn;
+  };
+
+  AppContainer.prototype.appendChild = function (child) {
+    this.root.appendChild(child, true);
+  };
+
+  AppContainer.prototype.removeChild = function (child) {
+    this.root.removeChild(child, true);
+  };
+
+  AppContainer.prototype.insertBefore = function (child, beforeChild) {
+    this.root.insertBefore(child, beforeChild, true);
+  };
+
+  return AppContainer;
+}();
+
+function fnBody(fn) {
+  return fn.toString().replace(/^[^{]*{\s*/, '').replace(/\s*}[^}]*$/, '');
+}
+
+function isClass(fn) {
+  if (typeof fn !== 'function') {
+    return false;
+  }
+
+  if (/^class[\s{]/.test(toString.call(fn))) {
+    return true;
+  } // babel.js classCallCheck() & inlined
+
+
+  var body = fnBody(fn);
+  return /classCallCheck/.test(body) || /TypeError\("Cannot call a class as a function"\)/.test(body);
+}
+
 function isClassComponent(Component) {
   return Component.prototype && typeof Component.prototype.render === 'function';
 }
@@ -1294,6 +1375,35 @@ function callbackName(name) {
   return 'on' + capitalize(name);
 }
 
+var context = {
+  lifecycleCallback: {},
+  registerLifecycle: function registerLifecycle(lifecycle, callback) {
+    var _this = this;
+
+    this.lifecycleCallback[lifecycle] = this.lifecycleCallback[lifecycle] || [];
+    this.lifecycleCallback[lifecycle].push(callback);
+    return function () {
+      _this.lifecycleCallback[lifecycle].splice(_this.lifecycleCallback[lifecycle].indexOf(callback), 1);
+    };
+  }
+};
+
+function _typeof(obj) {
+  "@babel/helpers - typeof";
+
+  if (typeof Symbol === "function" && typeof Symbol.iterator === "symbol") {
+    _typeof = function (obj) {
+      return typeof obj;
+    };
+  } else {
+    _typeof = function (obj) {
+      return obj && typeof Symbol === "function" && obj.constructor === Symbol && obj !== Symbol.prototype ? "symbol" : typeof obj;
+    };
+  }
+
+  return _typeof(obj);
+}
+
 /* eslint-disable */
 
 /* istanbul ignore file */
@@ -1310,9 +1420,43 @@ function callbackName(name) {
 // nor polyfill, then a plain number is used for performance.
 var hasSymbol = typeof Symbol === 'function' && Symbol["for"];
 var REACT_PORTAL_TYPE = hasSymbol ? Symbol["for"]('react.portal') : 0xeaca;
+var REACT_FRAGMENT_TYPE = hasSymbol ? Symbol["for"]('react.fragment') : 0xeacb;
+var REACT_STRICT_MODE_TYPE = hasSymbol ? Symbol["for"]('react.strict_mode') : 0xeacc;
+var REACT_PROFILER_TYPE = hasSymbol ? Symbol["for"]('react.profiler') : 0xead2;
+var REACT_PROVIDER_TYPE = hasSymbol ? Symbol["for"]('react.provider') : 0xeacd;
+var REACT_CONTEXT_TYPE = hasSymbol ? Symbol["for"]('react.context') : 0xeace; // TODO: We don't use AsyncMode or ConcurrentMode anymore. They were temporary
+var REACT_CONCURRENT_MODE_TYPE = hasSymbol ? Symbol["for"]('react.concurrent_mode') : 0xeacf;
 var REACT_FORWARD_REF_TYPE = hasSymbol ? Symbol["for"]('react.forward_ref') : 0xead0;
+var REACT_SUSPENSE_TYPE = hasSymbol ? Symbol["for"]('react.suspense') : 0xead1;
+var REACT_SUSPENSE_LIST_TYPE = hasSymbol ? Symbol["for"]('react.suspense_list') : 0xead8;
+var REACT_MEMO_TYPE = hasSymbol ? Symbol["for"]('react.memo') : 0xead3;
+var REACT_LAZY_TYPE = hasSymbol ? Symbol["for"]('react.lazy') : 0xead4;
+var REACT_FUNDAMENTAL_TYPE = hasSymbol ? Symbol["for"]('react.fundamental') : 0xead5;
+var REACT_RESPONDER_TYPE = hasSymbol ? Symbol["for"]('react.responder') : 0xead6;
+var REACT_SCOPE_TYPE = hasSymbol ? Symbol["for"]('react.scope') : 0xead7;
+function isValidElementType(type) {
+  return typeof type === 'string' || typeof type === 'function' || // Note: its typeof might be other than 'symbol' or 'number' if it's a polyfill.
+  type === REACT_FRAGMENT_TYPE || type === REACT_CONCURRENT_MODE_TYPE || type === REACT_PROFILER_TYPE || type === REACT_STRICT_MODE_TYPE || type === REACT_SUSPENSE_TYPE || type === REACT_SUSPENSE_LIST_TYPE || _typeof(type) === 'object' && type !== null && (type.$$typeof === REACT_LAZY_TYPE || type.$$typeof === REACT_MEMO_TYPE || type.$$typeof === REACT_PROVIDER_TYPE || type.$$typeof === REACT_CONTEXT_TYPE || type.$$typeof === REACT_FORWARD_REF_TYPE || type.$$typeof === REACT_FUNDAMENTAL_TYPE || type.$$typeof === REACT_RESPONDER_TYPE || type.$$typeof === REACT_SCOPE_TYPE);
+}
 var ForwardRef = REACT_FORWARD_REF_TYPE;
 var Portal = REACT_PORTAL_TYPE;
+
+function isPlainObject(obj) {
+  if (isValidElementType(obj)) {
+    return false;
+  }
+
+  if (_typeof(obj) == 'object' && obj !== null) {
+    if (typeof Object.getPrototypeOf == 'function') {
+      var proto = Object.getPrototypeOf(obj);
+      return proto === Object.prototype || proto === null;
+    }
+
+    return Object.prototype.toString.call(obj) == '[object Object]';
+  }
+
+  return false;
+}
 
 var __extends = undefined && undefined.__extends || function () {
   var _extendStatics = function extendStatics(d, b) {
@@ -1371,6 +1515,128 @@ function (_super) {
 
   return DefaultAppComponent;
 }(React.Component);
+
+function createAppConfig(App, extraConfig) {
+  var _this = this;
+
+  extraConfig = extraConfig || {};
+
+  var createConfig = function createConfig(AppComponent) {
+    var _a;
+
+    if (AppComponent === void 0) {
+      AppComponent = DefaultAppComponent;
+    }
+
+    var scope = {
+      _container: new AppContainer(_this),
+      _pages: [],
+      _instance: React.createRef(),
+      callLifecycle: function callLifecycle(lifecycle) {
+        var _a;
+
+        var args = [];
+
+        for (var _i = 1; _i < arguments.length; _i++) {
+          args[_i - 1] = arguments[_i];
+        }
+
+        var callbacks = context.lifecycleCallback[lifecycle] || [];
+        var result;
+        var callback = callbackName(lifecycle);
+
+        if (typeof extraConfig[callback] === 'function') {
+          return extraConfig[callback].apply(extraConfig, args);
+        }
+
+        callbacks.forEach(function (callback) {
+          result = callback.apply(void 0, args);
+        });
+
+        if (result) {
+          return result;
+        }
+
+        if (this._instance.current && this._instance.current[callback]) {
+          return (_a = this._instance.current)[callback].apply(_a, args);
+        }
+      },
+      _mount: function _mount(pageInstance) {
+        this._pages.push(pageInstance);
+
+        this._render();
+      },
+      _unmount: function _unmount(pageInstance) {
+        this._pages.splice(this._pages.indexOf(pageInstance), 1);
+
+        this._render();
+      },
+      _render: function _render() {
+        var props = {};
+
+        if (isClassComponent(AppComponent) || AppComponent.$$typeof === ForwardRef) {
+          props.ref = this._instance;
+        }
+
+        return render(React.createElement(AppComponent, props, this._pages.map(function (p) {
+          return p.element;
+        })), this._container);
+      }
+    };
+
+    var config = __assign$1(__assign$1((_a = {}, _a[CONFIG_SCOPE] = scope, _a), extraConfig), {
+      onLaunch: function onLaunch(options) {
+        this[CONFIG_SCOPE]._render();
+
+        this[CONFIG_SCOPE].callLifecycle(AppLifecycle.launch, options);
+      },
+      onShow: function onShow(options) {
+        this[CONFIG_SCOPE].callLifecycle(AppLifecycle.show, options);
+      },
+      onHide: function onHide() {
+        this[CONFIG_SCOPE].callLifecycle(AppLifecycle.hide);
+      },
+      onError: function onError(error) {
+        this[CONFIG_SCOPE].callLifecycle(AppLifecycle.error, error);
+      },
+      // 支付宝
+      onShareAppMessage: function onShareAppMessage(options) {
+        this[CONFIG_SCOPE].callLifecycle(AppLifecycle.shareAppMessage, options);
+      },
+      // 微信
+      onPageNotFound: function onPageNotFound(options) {
+        this[CONFIG_SCOPE].callLifecycle(AppLifecycle.pageNotFound, options);
+      }
+    });
+
+    return config;
+  };
+
+  if (isPlainObject(App)) {
+    var NativeApp =
+    /** @class */
+    function () {
+      function NativeApp() {
+        Object.assign(this, __assign$1({}, App));
+      }
+
+      return NativeApp;
+    }();
+
+    Object.assign(NativeApp.prototype, createConfig());
+    return new NativeApp();
+  } // 兼容老的写法和原生写法
+
+
+  if (isClass(App) && !isClassComponent(App)) {
+    // eslint-disable-next-line no-console
+    console.warn('使用非 React 组件定义 App 的方式已经废弃，详细请参考：https://remaxjs.org/guide/framework。');
+    Object.assign(App.prototype, createConfig());
+    return new App();
+  }
+
+  return createConfig(App);
+}
 
 var PageInstanceContext = React.createContext(null);
 
@@ -1856,6 +2122,22 @@ var __assign$2 = undefined && undefined.__assign || function () {
 
   return __assign$2.apply(this, arguments);
 };
+function createNativeComponent(name) {
+  return React.forwardRef(function (props, ref) {
+    if (typeof ref === 'string') {
+      console.error('不支持使用 string 获取小程序组件 ref，请使用回调或 React.createRef/React.useRef');
+    }
+
+    var newProps = __assign$2({}, props);
+
+    newProps.__ref = typeof ref === 'function' ? ref : function (e) {
+      if (ref) {
+        ref.current = e;
+      }
+    };
+    return React.createElement(name, newProps, props.children);
+  });
+}
 
 var unstable_batchedUpdates = ReactReconcilerInst.batchedUpdates;
 
@@ -2180,4 +2462,6 @@ var createVideoContext = my.createVideoContext;
 var getOpenUserInfo = promisify(my.getOpenUserInfo);
 
 exports.View = View;
+exports.createAppConfig = createAppConfig;
+exports.createNativeComponent = createNativeComponent;
 exports.createPageConfig = createPageConfig;
