--- packages/remax-cli/src/__tests__/fixtures/native/source/app.config.js
@@ -1,3 +1,0 @@
-module.exports = {
-  pages: ['pages/remax/index'],
-};
