+++ packages/remax-cli/src/__tests__/fixtures/native/source/components/native-component/index.axml
@@ -1,0 +1,1 @@
+<view>a</view>
