+++ packages/remax-cli/src/__tests__/fixtures/native/source/components/native-component/index.js
@@ -1,0 +1,3 @@
+Component({});
+
+export default {};
