+++ packages/remax-cli/src/__tests__/fixtures/native/source/components/native-component/index.json
@@ -1,0 +1,3 @@
+{
+  "component": true
+}
