+++ packages/remax-cli/src/__tests__/fixtures/native/source/native/pages/native/index.js
@@ -1,0 +1,1 @@
+Page({});
