+++ packages/remax-cli/src/__tests__/fixtures/native/source/pages/remax/index.tsx
@@ -1,6 +1,12 @@
 import * as React from 'react';
 import { View } from 'remax/alipay';
+import NativeComponent from '@/components/native-component';
 
 export default function RemaxPage(props: any) {
-  return <View>alipay view</View>;
+  return (
+    <View>
+      <NativeComponent />
+      alipay view
+    </View>
+  );
 }
