+++ packages/remax-cli/src/__tests__/fixtures/native/source/remaxApp.js
@@ -1,0 +1,3 @@
+export default function App(props) {
+  return props.children;
+}
