+++ packages/remax-cli/src/__tests__/fixtures/native/src/app.json
@@ -1,3 +1,8 @@
 {
-  "pages": ["pages/index/index", "pages/index2/index", "pages/remax/index"]
+  "pages": [
+    "pages/index/index",
+    "pages/index2/index",
+    "pages/remax/index",
+    "pages/native/index"
+  ]
 }
