+++ packages/remax-cli/src/__tests__/fixtures/native/src/pages/native/index.js
@@ -1,0 +1,1 @@
+Page({});
